// frontend/src/services/api/models/SalaryComponentCreate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { SalaryComponentType } from './SalaryComponentType';
export type SalaryComponentCreate = {
    name: string;
    type: SalaryComponentType;
    is_earning: boolean;
    description?: (string | null);
};