// frontend/src/services/api/models/SalaryComponentUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { SalaryComponentType } from './SalaryComponentType';
export type SalaryComponentUpdate = {
    name?: (string | null);
    type?: (SalaryComponentType | null);
    is_earning?: (boolean | null);
    description?: (string | null);
};